# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: LupexWallet API
  version: "1.0.0"
  description: >
    REST API однопользовательского приложения учета личных финансов.
    Технические идентификаторы — на английском, см. docs/architecture/ddd-model.md
    для доменной модели и docs/database/schema.md для схемы БД.
  license:
    name: Proprietary — внутренний проект, не публикуется
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /wallets/{walletId}:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    get:
      tags: [Wallets]
      summary: Получить кошелек
      operationId: getWallet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Wallets]
      summary: Изменить атрибуты кошелька (UC-02, без валюты)
      operationId: updateWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletUpdateRequest" }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wallet" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Wallets]
      summary: Удалить кошелек без истории (UC-04)
      operationId: deleteWallet
      responses:
        "204": { description: Удален }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Кошелек имеет историю операций/слепков (используйте /archive)
            или является основным.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /wallets/{walletId}/balance:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    get:
      tags: [BalanceHistory]
      summary: Баланс кошелька на дату (UC-18, UC-20)
      operationId: getWalletBalance
      description: >
        Без параметра date возвращает текущий баланс. Для даты раньше
        accounting_start_date возвращает сумму 0 (решено, Q13).
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceSnapshot" }
        "404": { $ref: "#/components/responses/NotFound" }
  /wallets/{walletId}/balance-history:
    parameters:
      - $ref: "#/components/parameters/WalletIdParam"
    get:
      tags: [BalanceHistory]
      summary: История слепков баланса кошелька за период
      operationId: listWalletBalanceHistory
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceSnapshotPage" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- Reference data ----------
...
  /exchange-rates/latest:
    get:
      tags: [ExchangeRates]
      summary: Последние известные курсы к текущей основной валюте (UC-08, UC-10)
      operationId: getLatestExchangeRates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LatestExchangeRates" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /exchange-rates/refresh:
    post:
      tags: [ExchangeRates]
      summary: Ручное обновление курсов (UC-09) — сбрасывает суточный таймер
      operationId: refreshExchangeRates
      responses:
        "200":
          description: Обновлено (или использован последний успешный курс при ошибке Frankfurter)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LatestExchangeRates" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502":
          description: Frankfurter недоступен; возвращены последние известные курсы.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LatestExchangeRates" }

  # ---------- Reporting ----------
...
  /audit-entries:
    get:
      tags: [Audit]
      summary: История изменений записи (UC-25)
      operationId: listAuditEntries
      parameters:
        - name: entityType
          in: query
          required: true
          schema: { type: string }
        - name: entityId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK, отсортировано по occurred_at по убыванию
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditEntryPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  parameters:
    WalletIdParam:
      name: walletId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CursorParam:
      name: cursor
      in: query
      description: Непрозрачный курсор для постраничной навигации.
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100, minimum: 1 }
  schemas:
...
    Money:
      type: object
      required: [amount, currencyId]
      properties:
        amount:
          type: string
          format: decimal
          description: Произвольная точность (ADR-0005); передается строкой, чтобы не терять точность в JSON-числах.
          example: "1234.56789012"
        currencyId: { type: string, format: uuid }
        currencyCode: { type: string, readOnly: true, example: "USD" }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    CursorPageMeta:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: [string, "null"] }
        hasMore: { type: boolean }

    # ---- Auth ----
...
    Wallet:
      type: object
      required:
        [id, name, walletTypeId, currencyId, initialBalance, accountingStartDate,
         currentBalance, includeInTotal, isPrimary, isArchived, displayOrder,
         createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string, maxLength: 200 }
        walletTypeId: { type: string, format: uuid }
        purposeDescription: { type: [string, "null"] }
        currencyId: { type: string, format: uuid }
        initialBalance: { $ref: "#/components/schemas/Money" }
        accountingStartDate: { type: string, format: date }
        currentBalance: { $ref: "#/components/schemas/Money", readOnly: true }
        includeInTotal: { type: boolean }
        isPrimary: { type: boolean, readOnly: true, description: "Меняется только через POST /wallets/{id}/set-primary" }
        isArchived: { type: boolean, readOnly: true }
        displayOrder: { type: integer }
        color: { type: [string, "null"] }
        icon: { type: [string, "null"] }
        createdAt: { type: string, format: date-time, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
...
    WalletUpdateRequest:
      type: object
      required: [name, walletTypeId, includeInTotal, displayOrder]
      properties:
        name: { type: string, maxLength: 200 }
        walletTypeId: { type: string, format: uuid }
        purposeDescription: { type: [string, "null"] }
        includeInTotal:
          type: boolean
          description: Игнорируется (остается true), если кошелек основной (Q8).
        displayOrder: { type: integer }
        color: { type: [string, "null"] }
        icon: { type: [string, "null"] }
...
    BalanceSnapshot:
      type: object
      required: [walletId, date, balance]
      properties:
        walletId: { type: string, format: uuid }
        date: { type: string, format: date }
        balance: { $ref: "#/components/schemas/Money" }
    BalanceSnapshotPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/BalanceSnapshot" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }

    # ---- Exchange rates ----
    ExchangeRateQuote:
      type: object
      required: [fromCurrencyId, toCurrencyId, rate]
      properties:
        fromCurrencyId: { type: string, format: uuid }
        toCurrencyId: { type: string, format: uuid }
        rate: { type: string, format: decimal }
    LatestExchangeRates:
      type: object
      required: [lastSuccessfulUpdate, rates]
      properties:
        lastSuccessfulUpdate:
          type: [string, "null"]
          format: date-time
          description: null, если ни одно обновление еще не было успешным.
        rates:
          type: array
          items: { $ref: "#/components/schemas/ExchangeRateQuote" }

    # ---- Reporting ----
...
    AuditFieldChange:
      type: object
      required: [field]
      properties:
        field: { type: string }
        oldValue: { type: [string, "null"] }
        newValue: { type: [string, "null"] }
    AuditEntry:
      type: object
      required: [id, entityType, entityId, action, occurredAt, actorKind]
      properties:
        id: { type: string, format: uuid }
        entityType: { type: string }
        entityId: { type: string, format: uuid }
        action: { type: string }
        occurredAt: { type: string, format: date-time }
        actorKind: { type: string, enum: [User, System] }
        actorSystemProcess: { type: [string, "null"] }
        changes:
          type: array
          items: { $ref: "#/components/schemas/AuditFieldChange" }
    AuditEntryPage:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/AuditEntry" } }
        pagination: { $ref: "#/components/schemas/CursorPageMeta" }
  responses:
    BadRequest:
      description: Некорректные параметры запроса
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Отсутствует или неверен токен доступа
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Запись не найдена
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: Конфликт с текущим состоянием (нарушение инварианта)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT